// src/API.ts
export type OrderInput = {
  orderCodeId: string;
  orderItems: Array<OrderItemInput>;
  orderStatus: OrderStatus;
};

export type OrderItemInput = {
  productId: string;
  variantId: string;
  optionIds: Array<string>;
  amount: number;
  extraText: string;
};

export enum OrderStatus {
  NEW = 'NEW',
  DONE = 'DONE',
  CANCELED = 'CANCELED',
}

export type Order = {
  __typename: 'Order';
  id: string;
  orderCodeId: string;
  orderItems: Array<OrderItem>;
  orderStatus: OrderStatus;
};

export type OrderItem = {
  __typename: 'OrderItem';
  productId: string;
  variantId: string;
  optionIds: Array<string>;
  amount: number;
  extraText: string;
};

export type OrderCodeInput = {
  deskId: string;
  isActive: boolean;
};

export type OrderCode = {
  __typename: 'OrderCode';
  id: string;
  deskId: string;
  isActive: boolean;
};